pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "top.py"
description = "A new, modern API wrapper for the top.gg API"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.9"
license = { text = "GNU LGPLv3" }
authors = [
    {name = "Nexus", email = "pip@nexy7574.co.uk"}
]
dynamic = ["version"]
classifiers =[
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Development Status :: 5 - Production/Stable"
]

[project.urls]
"Bug Tracker" = "https://github.com/nexy7574/top.py/issues"
"Documentation" = "https://toppy.readthedocs.io/"
"Source" = "https://github.com/nexy7574/top.py"


[project.optional-dependencies]
dev = [
    "discord-py",
    'pytest',
    'pytest-asyncio',
    'pytest-benchmark',
    'pytest-dependency',
    'ruff',
    "isort",
    "sphinx",
    "sphinx-rtd-dark-mode",
    "aioresponses"
]

[tool.black]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
